ACTIN-2087 PR feedback
