+++ algo/src/main/kotlin/com/hartwig/actin/algo/TrialMatcher.kt
@@ -1,13 +1,13 @@
 package com.hartwig.actin.algo
 
 import com.hartwig.actin.algo.evaluation.EvaluationFunctionFactory
-import com.hartwig.actin.algo.evaluation.EvaluationMessageCombiner
 import com.hartwig.actin.algo.evaluation.RuleMappingResources
 import com.hartwig.actin.algo.sort.CohortMatchComparator
 import com.hartwig.actin.algo.sort.TrialMatchComparator
 import com.hartwig.actin.datamodel.PatientRecord
 import com.hartwig.actin.datamodel.algo.CohortMatch
 import com.hartwig.actin.datamodel.algo.Evaluation
+import com.hartwig.actin.datamodel.algo.EvaluationMessage
 import com.hartwig.actin.datamodel.algo.EvaluationResult
 import com.hartwig.actin.datamodel.algo.TrialMatch
 import com.hartwig.actin.datamodel.trial.Eligibility
@@ -15,6 +15,20 @@ import com.hartwig.actin.datamodel.trial.EligibilityFunction
 import com.hartwig.actin.datamodel.trial.EligibilityRule
 import com.hartwig.actin.datamodel.trial.Trial
 import com.hartwig.actin.trial.sort.EligibilityComparator
+
+fun Evaluation.combineMessages(): Evaluation {
+    return copy(
+        passMessages = combineMessages(passMessages),
+        warnMessages = combineMessages(warnMessages),
+        undeterminedMessages = combineMessages(undeterminedMessages),
+        failMessages = combineMessages(failMessages)
+    )
+}
+
+private fun combineMessages(evaluations: Set<EvaluationMessage>): Set<EvaluationMessage> {
+    return evaluations.groupBy { it.combineBy() }
+        .mapValues { it.value.reduce { i, r -> i.combine(r) } }.values.toSet()
+}
 
 class TrialMatcher(private val evaluationFunctionFactory: EvaluationFunctionFactory) {
 
@@ -45,7 +59,7 @@ class TrialMatcher(private val evaluationFunctionFactory: EvaluationFunctionFact
 
     private fun evaluateEligibility(patient: PatientRecord, eligibility: List<Eligibility>): Map<Eligibility, Evaluation> {
         return eligibility.sortedWith(EligibilityComparator()).associateWith {
-            EvaluationMessageCombiner.combineMessages(evaluationFunctionFactory.create(it.function).evaluate(patient))
+            evaluationFunctionFactory.create(it.function).evaluate(patient).combineMessages()
         }
     }
